+++ CHANGELOG.md
@@ -6,7 +6,8 @@
   - Updated dependencies. `@babel/runtime` is now dep for all packages (PR #1246) (issue #964)
   - Fix group operator's custom jsonLogic function not being called (PR #1242) (issue #1241)
   - Fix import switch from JsonLogic (PR #1235) (issue #1230)
-  - Fix using "$not" as top-level operator in output to MongoDb (PR #1254) (issue #1250)
+  - Fix incorrect using of "$not" as top-level operator in output to MongoDb (PR #1254) (issue #1250)
+  - Fix using incorrect date format in output to MongoDb (PR #1254)
 - 6.6.14
   - Fixed import from JsonLogic when like op is used inside group with some group op (PR #1225) (issue #1221)
   - Support React 19 (PR #1229) (issue #1205)